apps/open-source/agent/requirements.txt
aiofiles==24.1.0
aiohttp==3.12.14
livekit-agents==1.2.5
livekit-plugins-deepgram==1.2.5
livekit-plugins-groq==1.2.5
livekit-plugins-silero==1.2.5
livekit-plugins-cartesia==1.2.5
livekit-plugins-openai==1.2.2
python-dotenv==1.1.1
sounddevice==0.5.2